Network weight returned by RPC call should be an integer.

+++ src/rpcmining.cpp
@@ -162,7 +162,7 @@ Value getmininginfo(const Array& params, bool fHelp)
     weight.push_back(Pair("total",   (uint64_t)nTotalWeight));
     obj.push_back(Pair("stakeweight", weight));
     obj.push_back(Pair("stakeinterest",  (uint64_t)COIN_YEAR_REWARD));
-    obj.push_back(Pair("netstakeweight", GetPoSVKernelPS()));
+    obj.push_back(Pair("netstakeweight", (uint64_t)GetPoSVKernelPS()));
     return obj;
 }
